+++ lib/flutter_supabase_chat_core.dart
@@ -6,4 +6,4 @@ export 'src/class/supabase_chat_core_config.dart';
 export 'src/class/user_online_status.dart';
 export 'src/util.dart';
 export 'src/widgets/user_online_status.dart';
-export 'src/widgets/user_status_detector.dart';
+export 'src/widgets/user_status_observer.dart';