+++ lib/src/class/supabase_chat_controller.dart
@@ -5,6 +5,8 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 
 import '../../flutter_supabase_chat_core.dart';
 
+/// Provides Supabase chat controller. Instance new class
+/// SupabaseChatController to manage a chat.
 class SupabaseChatController {
   late types.Room _room;
   final List<types.Message> _messages = [];
@@ -12,6 +14,9 @@ class SupabaseChatController {
   int _currentPage = 0;
   final _controller = StreamController<List<types.Message>>();
 
+  /// SupabaseChatController constructor
+  /// [pageSize] define a pagination size
+  /// [room] is required, is the controller's reference to the room
   SupabaseChatController({
     this.pageSize = 10,
     required types.Room room,
@@ -19,8 +24,10 @@ class SupabaseChatController {
     _room = room;
   }
 
+  /// return a SupabaseClient instance
   SupabaseClient get client => SupabaseChatCore.instance.client;
 
+  /// return a SupabaseChatCoreConfig instance
   SupabaseChatCoreConfig get config => SupabaseChatCore.instance.config;
 
   PostgrestTransformBuilder _query() => client
@@ -54,6 +61,10 @@ class SupabaseChatController {
     _controller.sink.add(_messages);
   }
 
+  /// Returns a stream of messages from Supabase for a specified room.
+  /// Only the amount of messages specified in [pageSize] will be loaded,
+  /// then it will be necessary to call the [loadPreviousMessages] method to get
+  /// the next page of messages
   Stream<List<types.Message>> get messages {
     _query().then((value) => _onData(value));
     client
@@ -72,6 +83,8 @@ class SupabaseChatController {
     return _controller.stream;
   }
 
+  /// This method allows to receive on the stream [messages] the next
+  /// page
   Future<void> loadPreviousMessages() async {
     _currentPage += 1;
     await _query().then((value) => _onData(value));